# Conflicts:
#	src/Application/ViewHandler.java

+++ src/Controllers/BoardGamesDetailPageController.java
@@ -16,6 +16,7 @@ import javafx.scene.layout.Region;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerException;
 import java.time.LocalDate;
+import java.util.Optional;
 
 public class BoardGamesDetailPageController implements Controller
 {
@@ -58,18 +59,28 @@ public class BoardGamesDetailPageController implements Controller
     min.setText(boardGame.getNumberOfPlayersMin() + "");
     max.setText(boardGame.getNumberOfPlayersMax() + "");
     numberOfVotes.setText(boardGame.getNumberOfVotes() + "");
-    comment.setDisable(true);
-    nameField.setDisable(true);
-    min.setDisable(true);
-    max.setDisable(true);
-    type.setDisable(true);
-    status.setDisable(true);
-    owner.setDisable(true);
-    numberOfVotes.setDisable(true);
+    comment.setMouseTransparent(true);
+    nameField.setMouseTransparent(true);
+    min.setMouseTransparent(true);
+    max.setMouseTransparent(true);
+    type.setMouseTransparent(true);
+    status.setMouseTransparent(true);
+    owner.setMouseTransparent(true);
+    numberOfVotes.setMouseTransparent(true);
+
+    comment.setFocusTraversable(false);
+    nameField.setFocusTraversable(false);
+    min.setFocusTraversable(false);
+    max.setFocusTraversable(false);
+    type.setFocusTraversable(false);
+    status.setFocusTraversable(false);
+    owner.setFocusTraversable(false);
+    numberOfVotes.setFocusTraversable(false);
     if (boardGame.getOwnerID() == 0)
     {
       owner.setText("Association");
-      selectAsocAsOwner.setDisable(true);
+      selectAsocAsOwner.setMouseTransparent(true);
+      selectAsocAsOwner.setFocusTraversable(false);
     }
     else
     {
@@ -98,13 +109,20 @@ public class BoardGamesDetailPageController implements Controller
   public void edit(ActionEvent actionEvent)
   {
 
-    comment.setDisable(false);
-    nameField.setDisable(false);
-    min.setDisable(false);
-    max.setDisable(false);
-    type.setDisable(false);
-    status.setDisable(false);
-    numberOfVotes.setDisable(false);
+    comment.setMouseTransparent(false);
+    nameField.setMouseTransparent(false);
+    min.setMouseTransparent(false);
+    max.setMouseTransparent(false);
+    type.setMouseTransparent(false);
+    status.setMouseTransparent(false);
+
+    comment.setFocusTraversable(true);
+    nameField.setFocusTraversable(true);
+    min.setFocusTraversable(true);
+    max.setFocusTraversable(true);
+    type.setFocusTraversable(true);
+    status.setFocusTraversable(true);
+    //numberOfVotes.setDisable(false);
     edit.setText("Save Changes");
     save = event -> {
       //System.out.println("hello");
@@ -116,30 +134,35 @@ public class BoardGamesDetailPageController implements Controller
       String typeOfGame = type.getValue().toString();
       String statusOfGame = status.getValue().toString();
       int ownerName = 1;
-      if (BoardGame.validateData(name, minimum, maximum))
-      {
-        BoardGame boardGame = new BoardGame(ID, name, typeOfGame, minimum,
-            maximum, statusOfGame, comments, ownerName, numberOfVotesOfGame);
-        model.setBoardGame(boardGame, ID);
-        try
+      try{
+        if (BoardGame.validateData(name, minimum, maximum))
         {
-          model.saveBoardGames();
-        }
-        catch (ParserConfigurationException e)
-        {
-          throw new RuntimeException(e);
-        }
-        catch (TransformerException e)
-        {
-          throw new RuntimeException(e);
-        }
-        setData();
-        edit.setOnAction(this::edit);
-        edit.setText("Edit");
-
+          BoardGame boardGame = new BoardGame(ID, name, typeOfGame, minimum,
+              maximum, statusOfGame, comments, ownerName, numberOfVotesOfGame);
+          model.setBoardGame(boardGame, ID);
+          try
+          {
+            model.saveBoardGames();
+          }
+          catch (ParserConfigurationException e)
+          {
+            throw new RuntimeException(e);
+          }
+          catch (TransformerException e)
+          {
+            throw new RuntimeException(e);
+          }
+          setData();
+          edit.setOnAction(this::edit);
+          edit.setText("Edit");
       }
-      else
-      {
+
+
+      } catch (Exception e){
+        Alert alert = new Alert(Alert.AlertType.ERROR);
+        alert.setTitle("Invalid Data");
+        alert.setHeaderText(e.getMessage());
+        Optional<ButtonType> result = alert.showAndWait();
       }
 
     };
@@ -149,10 +172,15 @@ public class BoardGamesDetailPageController implements Controller
 
   public void delete() throws ParserConfigurationException, TransformerException
   {
+    Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+    alert.setTitle("Confirmation");
+    alert.setHeaderText("Are you sure you want to delete the player ");
+    Optional<ButtonType> result = alert.showAndWait();
+    if(result.isPresent() && result.get() == ButtonType.OK){
     model.getBoardGamesList().deleteByID(ID);
     model.saveBoardGames();
     viewHandler.openView(3, -1);
-
+    }
   }
 
   public void goBack()
